% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacus.R
\name{abacus}
\alias{abacus}
\title{Abacus}
\usage{
abacus(
  nx = 30,
  ny = 100,
  max_size = 2,
  main_col = "black",
  bg_col = "white",
  s = 123
)
}
\arguments{
\item{nx}{Number of lines in x direction. Default 30.}

\item{ny}{Number of points per line. Default 100.}

\item{max_size}{Maximum size of points. Default 2.}

\item{main_col}{Vector of colours (or single colour). Default "black".}

\item{s}{Random seed. Default 123}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using points and lines.
}
