% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art.R
\name{art}
\alias{art}
\title{Heatmap Art}
\usage{
art(x_dim, y_dim)
}
\arguments{
\item{x_dim}{enter a random number for the x-dimension of your art}

\item{y_dim}{enter a random number for the y-dimension of your art}
}
\description{
With this function you can create random art with a heatmap
}
\details{
This is pretty cool!
}
