% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attraction.R
\name{attraction}
\alias{attraction}
\title{Attraction}
\usage{
attraction(
  n = 50000,
  a = -3,
  b = 1,
  c = 0.5,
  d = -1,
  main_col = "black",
  bg_col = "white"
)
}
\arguments{
\item{n}{Number of points. Default 50000.}

\item{a}{Parameter. Default -3.}

\item{b}{Parameter. Default 1.}

\item{c}{Parameter. Default 0.5.}

\item{d}{Parameter. Default -1.}

\item{main_col}{Colour of lines. Either a single
colour or one of the colour palettes from rcartocolor.
Default "black".}

\item{bg_col}{Background colour. Default "white".}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using strange attractors.
}
