% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_hole.R
\name{black_hole}
\alias{black_hole}
\title{Black Hole}
\usage{
black_hole(
  r_max = c(50, 150, 250, 350),
  n = 10000,
  lim = 300,
  col_palette = c("#FCDE9C", "#FAA476", "#F0746E", "#E34F6F", "#DC3977", "#B9257A",
    "#7C1D6F"),
  bg_col = "black",
  size = 0.01,
  a = 0.5,
  s = 1234
)
}
\arguments{
\item{r_max}{Vector of radii for the internal circle. Default \code{c(50, 150, 250, 350)}.}

\item{n}{Number of points per circle. Default 10000.}

\item{lim}{Numeric specifying size of grid. Default 400.}

\item{col_palette}{Vector of colours (or single colour). Default
\code{c("#FCDE9C", "#FAA476", "#F0746E", "#E34F6F", "#DC3977", "#B9257A", "#7C1D6F")}.}

\item{bg_col}{Background colour. Default "black".}

\item{size}{Size of points. Default 0.01.}

\item{a}{Transparency of points. Default 0.5.}

\item{s}{Random seed. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using points.
}
\examples{
black_hole()
}
