% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blending.R
\name{blending}
\alias{blending}
\title{Blending}
\usage{
blending(n = 100, down = "white", up = "black", s = 1234)
}
\arguments{
\item{n}{Number of point. Default 100.}

\item{down}{Colour to use on bottom. Default "white".}

\item{up}{Colour to use on top. Default "black".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object using a random walk.
}
\examples{
blending()
}
