% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxes}
\alias{boxes}
\title{Boxes}
\usage{
boxes(
  n = 100,
  perc = 0.1,
  col_palette = c("#D2FBD4", "#A5DBC2", "#7BBCB0", "#559C9E", "#3A7C89", "#235D72",
    "#123F5A"),
  bg_col = "black",
  s = 1234
)
}
\arguments{
\item{n}{Number of boxes}

\item{perc}{Relationship between box size and colour. Value between 0 and 1
where 0 represents randomness and 1 perfect identical. Default 0.1.}

\item{col_palette}{Vector of colours. Default \verb{"#D2FBD4" "#A5DBC2" "#7BBCB0" "#559C9E" "#3A7C89" "#235D72" "#123F5A"}.}

\item{bg_col}{Background colour. Default "black".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object from
treemaps.
}
\examples{
boxes()
}
