% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{bubbles}
\alias{bubbles}
\title{Bubbles}
\usage{
bubbles(
  num_circles = 20,
  main_col = "black",
  col_palette = rcartocolor::carto_pal(n = 12, "Bold"),
  bg_col = "white",
  s = 1234
)
}
\arguments{
\item{num_circles}{Number of circles. Default 20.}

\item{main_col}{Colour of non-highlighted rectangles. Default "black".}

\item{col_palette}{Vector of colours. Default "Bold" colour palette from rcartocolor. Must have 12 colours.}

\item{bg_col}{Background colour. Default "white".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object
}
\description{
This function generates a generative art ggplot object consisting of circles filled with ellipses.
}
\examples{
bubbles()
}
