% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullseye.R
\name{bullseye}
\alias{bullseye}
\title{Bullseye}
\usage{
bullseye(
  col_palette = c("#E01A4F", "#F15946", "#F9C22E", "#53B3CB", "#7DCFB6"),
  bg_col = "white",
  alpha = 0.3,
  s = 1234
)
}
\arguments{
\item{col_palette}{Colour palette. Default \code{c("#E01A4F", "#F15946", "#F9C22E", "#53B3CB", "#7DCFB6")}.}

\item{bg_col}{Background colour. Default \code{"white"}.}

\item{alpha}{Transparency. Default 0.3.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a layered generative art ggplot object using polar
bar charts.
}
\examples{
bullseye()
}
