% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos.R
\name{chaos}
\alias{chaos}
\title{Chaos}
\usage{
chaos(
  n_lines = 75,
  n_points = 10,
  n_circles = 20,
  line_col = "grey70",
  point_col = "black",
  circle_col = "white",
  circle_line_col = "black",
  bg_col = "white",
  min_circle = 0.01,
  max_circle = 0.1,
  linewidth = 0.2,
  alpha = 0.5,
  size = 0.3,
  s = 1234
)
}
\arguments{
\item{n_lines}{Number of lines. Default 75.}

\item{n_points}{Number of points. Default 10.}

\item{n_circles}{Number of circles. Default 20.}

\item{line_col}{Line colour. Default "grey70".}

\item{point_col}{Point colour. Default "black".}

\item{circle_col}{Circle fill colour. Default "white".}

\item{circle_line_col}{Circle line colour. Default "black".}

\item{bg_col}{Background colour. Default "white".}

\item{min_circle}{Minimum circle radius. Default 0.01.}

\item{max_circle}{Maximum circle radius. Default 0.1.}

\item{linewidth}{Linewidth of lines and circles. Default 0.2.}

\item{alpha}{Transparency of circles. Default 0.5.}

\item{size}{Size of points. Default 0.3.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object using a
lines, points, and circles.
}
\examples{
chaos()
}
