% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{circles}
\alias{circles}
\title{Circles}
\usage{
circles(
  n = 100,
  smoothness = 100,
  col_palette = rcartocolor::carto_pal(n = 12, "Bold"),
  line_col = NA,
  bg_col = "black",
  s = 1234
)
}
\arguments{
\item{n}{Number of nodes. Default 10.}

\item{smoothness}{Smoothness of lines on circles. Default 100.}

\item{col_palette}{Vector of colours. Default "Bold" colour palette from rcartocolor.}

\item{line_col}{Background colour. Default NA.}

\item{bg_col}{Background colour. Default "black".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using dendograms and circular graphs.
}
