% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.R
\name{circular}
\alias{circular}
\title{Generate Circular Art}
\usage{
circular(n = 100, main_col = "black", bg_col = "white", s = 56)
}
\arguments{
\item{n}{Number of steps from inside to outside. Default 100.}

\item{main_col}{Colour of lines. Default black.}

\item{bg_col}{Background colour. Default white.}

\item{s}{Seed value. Default 56.}
}
\value{
A ggplot object.
}
\description{
This function generates an abstract circular generative art ggplot object.
}
