% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected.R
\name{connected}
\alias{connected}
\title{Connected}
\usage{
connected(
  n = 100,
  n_geom = 2,
  random = FALSE,
  col_palette = RColorBrewer::brewer.pal(n = 9, "RdPu"),
  bg_col = "#ae217e",
  s = 1234
)
}
\arguments{
\item{n}{Number of lines to start. Default 100.}

\item{n_geom}{Number of points along path to create. Default 2.}

\item{random}{Boolean value for whether to randomise plotting order of edges.}

\item{col_palette}{Vector of colours. Default "RdPu" colour palette from
RColorBrewer.}

\item{bg_col}{Background colour. Default "#ae217e".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object connecting points
through arcs on a circle.
}
