% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawling.R
\name{crawling}
\alias{crawling}
\title{Crawling}
\usage{
crawling(
  n = 50,
  edge_colour = "black",
  node_size = 1,
  node_colour = "black",
  bg_col = "white",
  s = 1234
)
}
\arguments{
\item{n}{Number of nodes.}

\item{edge_colour}{Edge colour. Default "black".}

\item{node_size}{Node size. Default 1.}

\item{node_colour}{Node colour. Default "black".}

\item{bg_col}{Background colour. Default "white".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object using dendograms.
}
\examples{
crawling()
}
