% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criss_cross.R
\name{criss_cross}
\alias{criss_cross}
\title{Criss-cross}
\usage{
criss_cross(
  n = 25,
  bg_col = "white",
  line_col = "grey50",
  linewidth = 0.1,
  outline_col = "black",
  outline_width = 1.5,
  s = 1234
)
}
\arguments{
\item{n}{Number of lines per corner. Default 25.}

\item{bg_col}{Background colour. Default "white".}

\item{line_col}{Vector of colours (or single colour). Default "grey50".}

\item{linewidth}{Width of lines. Default 0.1.}

\item{outline_col}{Outline colour. Default "black".}

\item{outline_width}{Linewidth of outline. Default 1.5.}

\item{s}{Random seed. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using intersecting lines.
}
