% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosshatch.R
\name{crosshatch}
\alias{crosshatch}
\title{Crosshatch}
\usage{
crosshatch(
  n_x = 4,
  n_y = 4,
  n_lines = 10,
  line_overlap = 0.1,
  line_slope = 0.1,
  linewidth = 2,
  col_palette = c("#413C58", "#D1495B", "#EDAE49", "#00798C", "#003D5B"),
  bg_col = "#fafafa",
  interpolate = TRUE,
  s = 1234
)
}
\arguments{
\item{n_x}{Number of columns in grid. Default 4.}

\item{n_y}{Number of rows in grid. Default 4.}

\item{n_lines}{Number of lines per grid square. Default 10.}

\item{line_overlap}{Line overlap outside grid. Default 0.1.}

\item{line_slope}{Line slope within grid. Default 0.1.}

\item{linewidth}{Thickness of lines. Default 2.}

\item{col_palette}{Vector of colours. Default \code{c("#6497b1", "#6a359c", "#FFB04F", "#679c35", "#cd1076")}.}

\item{bg_col}{Background colour. Default \code{"gray10"}.}

\item{interpolate}{Boolean indicating if colours should be interpolated. Default TRUE.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using
overlapping lines in a grid.
}
\examples{
crosshatch()
}
