\name{higgins1990}
\docType{data}
\alias{data.higgins1990}
\title{Dataset to be used in the ART use example}
\description{
  This dataset was synthetically generated for a 3x3 completely randomized design, as shown in Table 1 of the paper:
  
  Higgins, J. J., Blair, R. C. and Tashtoush, S. (1990). The aligned rank transform procedure. Proceedings of the Conference on Applied Statistics in Agriculture. Manhattan, Kansas: Kansas State University, pp. 185-195.
  
  Columns Aligned_Row_Column and Ranks_Row_Column of the result correspond to the values of Table 3 of the same paper.
}
\usage{higgins1990}
\format{A table of 3 columns: two factors called Row and Column, and the Response variable. Loading this file gives access to the dataframe object \code{data.higgins1990}}
\source{File Higgins1990-Table1.csv of the ARTool zipped downloadable file: 
	\url{http://depts.washington.edu/aimgroup/proj/art/ARToolExe.zip}
}
\references{
  Higgins, J. J., Blair, R. C. and Tashtoush, S. (1990).
  The aligned rank transform procedure. Proceedings of the
  Conference on Applied Statistics in Agriculture.
  Manhattan, Kansas: Kansas State University, pp. 185-195.

  Wobbrock, J.O., Findlater, L., Gergle, D. and Higgins,
  J.J. (2011). The Aligned Rank Transform for nonparametric
  factorial analyses using only ANOVA procedures.
  Proceedings of the ACM Conference on Human Factors in
  Computing Systems (CHI '11). New York: ACM Press, pp.
  143-146.

  ARTool: free visual software tool implementing ART,
  written in the C# language :
  \url{http://depts.washington.edu/aimgroup/proj/art/}
}
\keyword{datasets}
