% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distort.R
\name{distort}
\alias{distort}
\title{Distort}
\usage{
distort(
  n_x = 5,
  n_y = 5,
  deg_jitter = 0.4,
  col_palette = c("#4D4D4D", "#888888", "#AEAEAE", "#CCCCCC", "#E6E6E6"),
  line_col = "transparent",
  linewidth = 1,
  bg_col = "#333333",
  s = 1234
)
}
\arguments{
\item{n_x}{Number of columns. Default 4.}

\item{n_y}{Number of rows. Default 4.}

\item{deg_jitter}{Degree of jitter. Default 0.4.}

\item{col_palette}{Vector of colours. Must be at least length 4.
Default \code{c("#4D4D4D", "#888888", "#AEAEAE", "#CCCCCC", "#E6E6E6")}}

\item{line_col}{Colour of lines. Default \code{"transparent"}.}

\item{linewidth}{Width of lines. Default 1.}

\item{bg_col}{Background colour. Default \code{"#333333"}.}

\item{s}{Random seed. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using polygons
}
\examples{
distort()
}
