% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Divide}
\usage{
divide(
  num_lines = 30,
  col_palette = PrettyCols::prettycols("TangerineBlues"),
  s = 1234
)
}
\arguments{
\item{num_lines}{Number of intersecting lines. Default 30.}

\item{col_palette}{Vector of colours. Default PrettyCols::prettycols("TangerineBlues").}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object from intersecting lines.
}
