% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots}
\alias{dots}
\title{Dots}
\usage{
dots(
  n_x = 50,
  n_y = 100,
  jitter_size_width = 0.5,
  jitter_size_height = 0.5,
  col_palette = "Purp",
  bg_col = "#63589f",
  s = 1234
)
}
\arguments{
\item{n_x}{Number of rotational points. Default 50.}

\item{n_y}{Number of outwards points. Default 100.}

\item{jitter_size_width}{Size of jitter width. Default 0.5.}

\item{jitter_size_height}{Size of jitter height. Default 0.5.}

\item{col_palette}{Colour palette from rcartocolor. Default "Purp".}

\item{bg_col}{Background colour. Default "#63589f".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using polar coordinates.
}
