% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fading.R
\name{fading}
\alias{fading}
\title{Fading}
\usage{
fading(
  n_layers = 6,
  n_points = 10,
  col_palette = c("#FCDE9C", "#FAA476", "#F0746E", "#E34F6F", "#DC3977", "#B9257A",
    "#7C1D6F"),
  s = 1234
)
}
\arguments{
\item{n_layers}{Number of layers. Default 6.}

\item{n_points}{Number of points per layer area. Default 10.}

\item{col_palette}{Vector of colours. Default \verb{c("#FCDE9C", "#FAA476", "#F0746E", "#E34F6F", "#DC3977", "#B9257A", "#7C1D6F"}.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using
voronoi tiles.
}
\examples{
fading()
}
