% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_fields.R
\name{flow_fields}
\alias{flow_fields}
\title{Flow fields}
\usage{
flow_fields(
  n = 10000,
  granularity = 1000,
  x_freq = 1,
  y_freq = 1,
  alpha = 1,
  line_col = c("#edf8fb", "#bfd3e6", "#9ebcda", "#8c96c6", "#8c6bb1", "#88419d",
    "#6e016b"),
  bg_col = "lightgrey",
  s = 1234
)
}
\arguments{
\item{n}{Number of lines. Default 10000.}

\item{granularity}{How fine to draw the grid. Default 1000.}

\item{x_freq}{Frequency of x simplex noise. Default 1.}

\item{y_freq}{Frequency of y simplex noise. Default 1.}

\item{alpha}{Transparency of lines. Default 1.}

\item{line_col}{Line colours. Vector (or single element) of colours. Default c("#edf8fb","#bfd3e6","#9ebcda","#8c96c6","#8c6bb1","#88419d","#6e016b")}

\item{bg_col}{Background colour. Default "white".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object from using particle traces.
See also https://www.williamrchase.com/post/flow-fields-12-months-of-art-september/
}
