% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractals.R
\name{fractals}
\alias{fractals}
\title{Fractals}
\usage{
fractals(
  N = 25,
  col_palette = MetBrewer::met.brewer("Demuth", n = 25),
  shift = 0,
  left = -1,
  right = 1,
  y_param = 3,
  resolution = 0.005,
  dist_max = 4
)
}
\arguments{
\item{N}{Number of iterations. Default 25.}

\item{col_palette}{Vector of colours of length >= N. Default MetBrewer::met.brewer("Demuth", n = 25).}

\item{shift}{Offset of y-values. Default 0.}

\item{left}{Start rangle of x-axis. Default -1.}

\item{right}{End range of x-axis. Default 1.}

\item{y_param}{Rate of y growth. Default 3.}

\item{resolution}{Resolution of grid. Default 0.005.}

\item{dist_max}{Size of center area. Default 4.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using fractal patterns. Inspired by https://www.r-bloggers.com/2010/08/fractals-in-r/
}
