% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{gradients}
\alias{gradients}
\title{Generates generative art as a grid of gradient colour fades}
\usage{
gradients(nx = 4, ny = 4, bg_col = "white", linewidth = 1.5, s = 1234)
}
\arguments{
\item{nx}{Number of columns. Default 4.}

\item{ny}{Number of rows. Default 4.}

\item{bg_col}{Background colour. Default "white".}

\item{linewidth}{Linewidth. Default 1.5.}

\item{s}{Random seed. Default 1234.}
}
\value{
Tibble
}
\description{
Generates generative art as a grid of gradient colour fades
}
\examples{
gradients()
}
