% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imploding_hexagon.R
\name{imploding_hexagon}
\alias{imploding_hexagon}
\title{Imploding Hexagon}
\usage{
imploding_hexagon(
  n = 25000,
  size = 0.05,
  bg_col = "grey10",
  col_palette = grDevices::grey.colors(n = 10, start = 0.1, end = 0.9),
  random = TRUE,
  implode_factor = 0.5,
  s = 1234
)
}
\arguments{
\item{n}{Number of points. Default 25000.}

\item{size}{Size of points. Default 0.1.}

\item{bg_col}{Background colour. Default "grey10".}

\item{col_palette}{Colour palette. Default
\code{grDevices::grey.colors(n = 10, start = 0.1, end = 0.7)}.}

\item{random}{Boolean. Should colours be arranged randomly or not?
Default \code{TRUE}.}

\item{implode_factor}{Image implode factor. Default 0.5.}

\item{s}{Random seed. Default 123.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using spatial hexagons and Magick.
}
\examples{
imploding_hexagon()
}
