% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squiggles.R
\name{make_lines}
\alias{make_lines}
\title{Make lines}
\usage{
make_lines(res = 100)
}
\arguments{
\item{res}{Resolution of grid. Default 100.}
}
\value{
A numeric vector
}
\description{
This function generates smooth random walks from a normal distribution.
}
