% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{make_square}
\alias{make_square}
\title{Generates a tibble of line segments and colours}
\usage{
make_square(xmin, xmax, ymin, ymax, low, high)
}
\arguments{
\item{xmin}{x-min coord}

\item{xmax}{x-max coord}

\item{ymin}{y-min coord}

\item{ymax}{y-max coord}

\item{low}{Hex value for left colour}

\item{high}{Hex value for right colour}
}
\value{
Tibble
}
\description{
Generates a tibble of line segments and colours
}
