% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{mesh}
\alias{mesh}
\title{Mesh}
\usage{
mesh(
  n_x = 10,
  n_y = 10,
  col_palette = c("#0e7c7b", "#17bebb", "#d4f4dd", "#d62246", "#4b1d3f"),
  linewidth = 2,
  rayshade = FALSE
)
}
\arguments{
\item{n_x}{Number of squares per row. Default 10.}

\item{n_y}{Number of squares per column. Default 10.}

\item{col_palette}{Vector of colours. Default
\code{c("#0e7c7b", "#17bebb", "#d4f4dd", "#d62246", "#4b1d3f")}.}

\item{linewidth}{Width of borders around squares. Default 2.}

\item{rayshade}{Boolean determining whether the returned plot should be
converted to three dimensional using rayshader. If \code{TRUE}, \code{{rayshader}}
is required to be installed. Default \code{FALSE}.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using rectangles.
}
\examples{
mesh()
}
