% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrored.R
\name{mirrored}
\alias{mirrored}
\title{Mirrored}
\usage{
mirrored(
  n = 15,
  w = 4,
  col_palette = c("#420f75", "#7640a9", "#ad72d6", "#e7a8fb", "#f5f5f5", "#f8b150",
    "#c17d17", "#8a4d00", "#552000"),
  s = 1234
)
}
\arguments{
\item{n}{Number of boxes per quadrant. Default 15.}

\item{w}{Weighting towards first colour of palette. Minimum of 2. Default 4.}

\item{col_palette}{Vector of colours. Default \code{c("#420f75", "#7640a9", "#ad72d6", "#e7a8fb", "#f5f5f5", "#f8b150", "#c17d17", "#8a4d00", "#552000")}.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object from rectangles.
}
\examples{
mirrored()
}
