% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moire.R
\name{moire}
\alias{moire}
\title{Moiré}
\usage{
moire(
  inner_n = 20,
  dist = 10,
  inner_col = "grey40",
  outer_col = "grey60",
  bg_col = "grey10",
  inner_r = 0.5,
  outer_r = 0.2
)
}
\arguments{
\item{inner_n}{Number of bottom left circles. Default 20.}

\item{dist}{Distance between bottom left and top right blocks. Default 4.}

\item{inner_col}{Colour of bottom left circles. Default "grey40".}

\item{outer_col}{Colour of top right circles. Default "grey60".}

\item{bg_col}{Background colour. Default "grey10".}

\item{inner_r}{Radius of bottom left circles. Default 0.5.}

\item{outer_r}{Radius of top right circles. Default 0.2.}
}
\value{
A ggplot object.
}
\description{
This function creates generative art inspired by the Moiré effect.
}
\examples{
moire()
}
