% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic}
\alias{mosaic}
\title{Mosiac}
\usage{
mosaic(
  n = 100,
  fill_cols = c("#4B3F72", "#CBB3BF", "#FFC857", "#119DA4", "#19647E"),
  line_col = "white",
  bg_col = "white",
  line_size = 1,
  x_means = c(0, 10, 5),
  y_means = c(0, 7, 8),
  xy_var = 2,
  s = 1234
)
}
\arguments{
\item{n}{Number of points to generate tiles from. Default 100.}

\item{fill_cols}{Vector of colours to fill tiles with, Default c("#4B3F72", "#CBB3BF", "#FFC857", "#119DA4", "#19647E").}

\item{line_col}{Colour of lines between tiles, Default "white".}

\item{bg_col}{Background colour. Default "white".}

\item{line_size}{Thickness of lines between tiles. Default 1.}

\item{x_means}{Vector of any number of means for the x-coordinate. Default c(0, 10, 5).}

\item{y_means}{Vector of any number of means for the y-coordinate. Default c(0, 7, 8).}

\item{xy_var}{Numeric varaince of x and y points. Default 2.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object from voronoi tiles.
}
