% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpendicular.R
\name{perpendicular}
\alias{perpendicular}
\title{Perpendicular}
\usage{
perpendicular(
  n = 100,
  max_length = 7,
  linewidth = 0.5,
  main_col = "black",
  bg_col = "white",
  s = 123
)
}
\arguments{
\item{n}{Number of rectangles. Default 100.}

\item{max_length}{Maximum length of line. Default 7.}

\item{linewidth}{Line width of line. Default 0.5.}

\item{main_col}{Colour of lines. Default "black".}

\item{bg_col}{Background colour. Default "white".}

\item{s}{Seed value. Default 123.}
}
\value{
A ggplot object
}
\description{
This function generates a generative art ggplot object featuring multiple coloured perpendicular lines.
}
