% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzles.R
\name{puzzles}
\alias{puzzles}
\title{Puzzles}
\usage{
puzzles(
  n = 200,
  num_groups = 30,
  col_palette = PrettyCols::prettycols("Beach"),
  bg_col = "white",
  s = 1234
)
}
\arguments{
\item{n}{Number of boxes. Default 200.}

\item{num_groups}{Number of larger boxes. Default 30.}

\item{col_palette}{Vector of colours. Default \code{PrettyCols::prettycols("Beach")}.}

\item{bg_col}{Background colour. Default \code{"white"}.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object from treemaps.
}
\examples{
puzzles()
}
