% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_tessellation.R
\name{random_tessellation}
\alias{random_tessellation}
\title{Random Tessellation}
\usage{
random_tessellation(
  n_x = 10,
  n_y = 10,
  deg_jitter = 0.1,
  linewidth = 0.5,
  line_col = "black",
  bg_col = "black",
  col_palette = PrettyCols::prettycols("Lively"),
  s = 1234
)
}
\arguments{
\item{n_x}{Number of polygons per row. Default 10.}

\item{n_y}{Number of polygons per column. Default 10.}

\item{deg_jitter}{Numeric between 0 and 0.5 specifying the degree of jitter. Default 0.1.}

\item{linewidth}{Width of lines between polygons. Default 0.5.}

\item{line_col}{Colour of lines between polygons. Default \code{"black"}.}

\item{bg_col}{Background colour. Default \code{"black"}.}

\item{col_palette}{Vector of colours. Can be any length. Default \code{PrettyCols::prettycols("Lively")}.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using polygons.
}
\examples{
random_tessellation()
}
