% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{rectangles}
\alias{rectangles}
\title{Rectangles}
\usage{
rectangles(
  n = 100,
  max_height = 7,
  max_width = 5,
  size = 2,
  main_col = "lightgrey",
  col_palette = c("#80BA5A", "#E68310", "#008695", "#CF1C90", "#F97B72", "#4B4B8F"),
  bg_col = "white",
  weight = 0.25,
  s = 1234
)
}
\arguments{
\item{n}{Number of rectangles. Default 100.}

\item{max_height}{Maximum height of rectangle. Default 7.}

\item{max_width}{Maximum width of rectangle. Default 5.}

\item{size}{Line width of rectangles. Default 2.}

\item{main_col}{Colour of non-highlighted rectangles. Default \code{"lightgrey"}.}

\item{col_palette}{Vector of colours. Default \code{c("#80BA5A", "#E68310", "#008695", "#CF1C90", "#F97B72", "#4B4B8F")}.}

\item{bg_col}{Background colour. Default \code{"white"}.}

\item{weight}{Sampling weight. Default 0.25.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object
}
\description{
This function generates a generative art ggplot object featuring multiple
coloured rectangles.
}
\examples{
rectangles()
}
