% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riley.R
\name{riley}
\alias{riley}
\title{Riley}
\usage{
riley(n_x = 9, n_y = 9, offset = 3, main_col = "black", bg_col = "white")
}
\arguments{
\item{n_x}{Number of circles in x-direction. Default 9.}

\item{n_y}{Number of circles in y-direction. Default 9.}

\item{offset}{Numeric giving offset of y-axis. Default 3.}

\item{main_col}{Main colour. Default "black".}

\item{bg_col}{Background colour. Default "white".}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object from
intersecting lines.
Inspired by Balm by Bridget Riley.
}
