% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rings.R
\name{rings}
\alias{rings}
\title{Rings}
\usage{
rings(
  col_palette = PrettyCols::prettycols("Lively"),
  bg_col = "#343046",
  x_ring = c(0.2, 0.9),
  y_ring = c(0.2, 1.8),
  r_ring = c(0.6, 0.4),
  x0 = c(0, 1),
  y0 = c(0, 2),
  r = c(1, 0.7),
  n = c(80, 80),
  s = 1234
)
}
\arguments{
\item{col_palette}{Vector of colours. Default PrettyCols::prettycols("Lively").}

\item{bg_col}{Background colour. Default "#343046".}

\item{x_ring}{Vector of x-co-ordinates for centre of gaps. Default c(0.2, 0.9).}

\item{y_ring}{Vector of y-co-ordinates for centre of gaps. Default c(0.2, 1.8).}

\item{r_ring}{Vector of radii for centre of gaps. Default c(0.6, 0.4).}

\item{x0}{Vector of x-co-ordinates for pie charts. Default c(0, 1).}

\item{y0}{Vector of y-co-ordinates for pie charts. Default c(0, 2).}

\item{r}{Vector of radii for pie charts. Default c(1, 0.7).}

\item{n}{Vector of slices per pie chart. Default c(80, 80).}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object using pie charts
and circles.
}
