% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riso_circles.R
\name{riso_circles}
\alias{riso_circles}
\title{Riso Circles}
\usage{
riso_circles(
  n_x = 4,
  n_y = 4,
  n_circles = 2,
  r = 0.5,
  jitter_x = 0.25,
  jitter_y = 0.3,
  alpha = 0.5,
  col_palette = c("#6497b1", "#6a359c", "#FFB04F", "#679c35", "#cd1076"),
  circle_bg = "#fafafa",
  bg_col = "gray10",
  interpolate = TRUE,
  s = 1234
)
}
\arguments{
\item{n_x}{Number of columns in grid. Default 4.}

\item{n_y}{Number of rows in grid. Default 4.}

\item{n_circles}{Number of circles per grid square. Default 2.}

\item{r}{Radius of circles. Default 0.5.}

\item{jitter_x}{Jitter in x direction for circle positions. Default 0.25.}

\item{jitter_y}{Jitter in y direction for circle positions. Default 0.3.}

\item{alpha}{Transparency of circles. Default 0.5.}

\item{col_palette}{Vector of colours. Default \code{c("#6497b1", "#6a359c", "#FFB04F", "#679c35", "#cd1076")}.}

\item{bg_col}{Background colour. Default "gray10".}

\item{interpolate}{Boolean indicating if colours should be interpolated. Default TRUE.}

\item{s}{Seed value. Default 1234.}

\item{cicrle_bg}{Background colour of circles. Default "#fafafa".}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using
overlapping semi-transparent circles.
}
