% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shells.R
\name{shells}
\alias{shells}
\title{Shells}
\usage{
shells(n = 4, alpha = 1, main_col = "black", bg_col = "white")
}
\arguments{
\item{n}{Number of spirals. Default 4.}

\item{alpha}{Transparency of lines. Default 1.}

\item{main_col}{Colour scheme of art. Default "black".}

\item{bg_col}{Background colour. Default "white".}
}
\value{
A ggplot object.
}
\description{
This function generates a layered generative art ggplot object sine and cosine waves and lines.
}
