% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirals.R
\name{spirals}
\alias{spirals}
\title{Spirals}
\usage{
spirals(perc = 0.2, s = 1234)
}
\arguments{
\item{perc}{Percentage of data points to be non-NA. Default 0.2.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object consisting of dots arranged in a spiral,
}
