% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_grid.R
\name{split_grid}
\alias{split_grid}
\title{Split grid}
\usage{
split_grid(
  n_x = 4,
  n_y = 4,
  col_palette = c("#FF8811", "#9DD9D2", "#046E8F", "#D44D5C"),
  grid_col = "white",
  grid_width = 1,
  s = 1234
)
}
\arguments{
\item{n_x}{Number of columns. Default 4.}

\item{n_y}{Number of rows. Default 4.}

\item{col_palette}{Vector of colours. Must be at least length 4.
Default \code{c("#FF8811", "#9DD9D2", "#046E8F", "#D44D5C")}}

\item{grid_col}{Colour of grid lines. Default "white".}

\item{grid_width}{Linewidth of grid lines. Default 1.}

\item{s}{Random seed. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using polygons
}
