% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_jitter.R
\name{split_jitter}
\alias{split_jitter}
\title{Split jitter}
\usage{
split_jitter(
  n_x = 5,
  n_y = 5,
  deg_jitter = 0.4,
  col_palette = c("#416322", "#4e7628", "#5a892f", "#679c35", "#74af3b", "#80c044",
    "#8dc657"),
  line_col = "transparent",
  linewidth = 1,
  bg_col = "#679c35",
  s = 1234
)
}
\arguments{
\item{n_x}{Number of columns. Default 4.}

\item{n_y}{Number of rows. Default 4.}

\item{deg_jitter}{Degree of jitter. Default 0.4.}

\item{col_palette}{Vector of colours. Must be at least length 4.
Default \code{c("#416322", "#4e7628", "#5a892f", "#679c35", "#74af3b", "#80c044", "#8dc657")}}

\item{line_col}{Colour of lines. Default \code{"transparent"}.}

\item{linewidth}{Width of lines. Default 1.}

\item{bg_col}{Background colour. Default \code{"#679c35"}.}

\item{s}{Random seed. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using polygons
}
\examples{
split_jitter()
}
