% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squares.R
\name{squares}
\alias{squares}
\title{Squares}
\usage{
squares(
  n = 7,
  line_col = "white",
  pattern_col = "white",
  pattern_fill = "black",
  pattern_size = 0.4,
  size = 1.5,
  s = 1234
)
}
\arguments{
\item{n}{Number of squares per row. Default 7}

\item{line_col}{Colour of lines between squares. Default "white".}

\item{pattern_col}{Colour of pattern lines. Default "white".}

\item{pattern_fill}{Colour of pattern background. Default "black".}

\item{pattern_size}{Size of pattern. Default 0.4.}

\item{size}{Size of lines between squares. Default 1.5.}

\item{s}{Random seed. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using pattern-filled squares
}
\examples{
squares()
}
