% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squiggles.R
\name{squiggles}
\alias{squiggles}
\title{Squiggles}
\usage{
squiggles(
  res = 100,
  num_lines = 100,
  perc = 0.1,
  alpha_low = 0.5,
  alpha_high = 1,
  line_col = "white",
  bg_col = "black",
  s = 1234
)
}
\arguments{
\item{res}{Resolution of grid. Default 100.}

\item{num_lines}{Number of lines to draw. Default 100.}

\item{perc}{Percentage to colour darker. Default 0.1.}

\item{alpha_low}{Transparency of majority of lines. Default 0.5.}

\item{alpha_high}{Transparency of minority of lines. Default 1.}

\item{line_col}{Colour of lines. Default "white".}

\item{bg_col}{Background colour. Default "black".}

\item{s}{Random seed. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates generative art from multiple smooth lines.
}
