% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackture.R
\name{stack}
\alias{stack}
\title{Stack}
\usage{
stack(x0, y0, min_height, max_height, min_width, max_width)
}
\arguments{
\item{x0}{x-coordinate of middle of rectangle.}

\item{y0}{y-coordinate of middle of rectangle.}

\item{min_height}{minimum height.}

\item{max_height}{maximum height.}

\item{min_width}{minimum width.}

\item{max_width}{maximum width.}
}
\value{
An sf object.
}
\description{
This function generates an sf object.
}
