% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackture.R
\name{stackture}
\alias{stackture}
\title{Stackture}
\usage{
stackture(
  n_x = 8,
  n_y = 8,
  min_height = 1,
  max_height = 1.5,
  min_width = 1,
  max_width = 1.5,
  interpolate = TRUE,
  col_palette = c("#A053A1", "#DB778F", "#E69F52", "#09A39A", "#5869C7"),
  bg_col = "#004B67",
  s = 1234
)
}
\arguments{
\item{n_x}{Number of columns in grid. Default 8.}

\item{n_y}{Number of rows in grid. Default 8.}

\item{min_height}{minimum height.}

\item{max_height}{maximum height.}

\item{min_width}{minimum width.}

\item{max_width}{maximum width.}

\item{interpolate}{Boolean indicating if colours should be interpolated. Default TRUE.}

\item{col_palette}{Vector of colours. Default \code{c("#A053A1", "#DB778F", "#E69F52", "#09A39A", "#5869C7")}.}

\item{bg_col}{Background colour. Default "#004B67".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using
overlapping semi-transparent circles.
}
