% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{streams}
\alias{streams}
\title{Streams}
\usage{
streams(
  bg_col = "white",
  line_col = "white",
  fill_col = c("#5F4690", "#1D6996", "#38A6A5", "#0F8554", "#73AF48", "#EDAD08",
    "#E17C05", "#CC503E", "#94346E", "#6F4070"),
  type = "right",
  s = 1234
)
}
\arguments{
\item{bg_col}{Background colour. Default "white".}

\item{line_col}{Line colour. Default "white".}

\item{fill_col}{Vector of fill colours.}

\item{type}{Rotation of stream. Default "right".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object from stream charts.
}
\examples{
streams()
}
