% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripes.R
\name{stripes}
\alias{stripes}
\title{Generate Striped Art}
\usage{
stripes(perc, n = 3, s = 1234)
}
\arguments{
\item{perc}{Percentage of data points to be sorted}

\item{n}{Number of rows}

\item{s}{Seed value}
}
\value{
A ggplot object
}
\description{
This function generates a generative art ggplot object featuring rows of stripes.
}
