% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunbursts.R
\name{sunbursts}
\alias{sunbursts}
\title{Sunbursts}
\usage{
sunbursts(
  n = 100,
  x_means = c(0, 10, 5),
  y_means = c(0, 7, 8),
  xy_var = 5,
  low = "#074050",
  high = "#d3f2a3",
  s = 1234
)
}
\arguments{
\item{n}{Granularity. Default 100.}

\item{x_means}{Vector of any number of means for the x-coordinate. Default c(0, 10, 5).}

\item{y_means}{Vector of any number of means for the y-coordinate. Default c(0, 7, 8).}

\item{xy_var}{Numeric varaince of x and y points. Default 5.}

\item{low}{Colour of background. Default "#074050".}

\item{high}{Colour of sunburst points. Default "#d3f2a3".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object from 2d density plots
}
