% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunsets.R
\name{sunsets}
\alias{sunsets}
\title{Sunsets}
\usage{
sunsets(
  num_bars = 6,
  n = 500,
  col_palette = c("#413C58", "#D1495B", "#EDAE49", "#00798C", "#003D5B"),
  bg_col = "#413C58",
  vertical = FALSE,
  fade_vertical = FALSE,
  alpha = 1,
  s = 1234
)
}
\arguments{
\item{num_bars}{Number of bars. Default 6.}

\item{n}{Number of lines per bar. Default 500.}

\item{col_palette}{Colour palette. Default \code{c("#413C58", "#D1495B", "#EDAE49", "#00798C", "#003D5B")}.}

\item{bg_col}{Background colour. Default \code{"#413C58"}.}

\item{vertical}{Boolean indicating whether bars should be
vertical. Default FALSE.}

\item{fade_vertical}{Boolean indicating whether the colouring
should be vertical. Default FALSE.}

\item{alpha}{Transparency of coloured bars. Default 1.}

\item{s}{Random seed. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function creates generative art using faceted segments.
}
\examples{
sunsets()
}
