% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aRt.R
\docType{data}
\name{theme_aRt}
\alias{theme_aRt}
\title{aRt theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 97.
}
\usage{
theme_aRt
}
\description{
Custom ggplot2 theme for aRt objects
}
\keyword{datasets}
