% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{tiles}
\alias{tiles}
\title{Tiles}
\usage{
tiles(
  n_x = 12,
  n_y = 12,
  col_palette = c("#67322e", "#99610a", "#6e948c", "#2c6b67", "#122c43"),
  rayshade = FALSE,
  shadow_intensity = 0.5,
  sunangle = 315,
  s = 1234
)
}
\arguments{
\item{n_x}{Number of polygons per row. Default 12.}

\item{n_y}{Number of polygons per column. Default 18.}

\item{col_palette}{Vector of colours. Default \code{c("#67322e", "#99610a", "#6e948c", "#2c6b67", "#122c43")} which is the Veronese palette from MetBrewer.}

\item{rayshade}{Boolean determining whether the returned plot should
be converted to three dimensional using rayshader. If \code{TRUE}, \code{{rayshader}}
is required to be installed. Default \code{FALSE}.}

\item{shadow_intensity}{Intensity of shading for 3D elements, Default 0.5.}

\item{sunangle}{Angle of the sun. Default 315.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using square polygons.
}
\examples{
tiles()
}
