% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wander.R
\name{wander}
\alias{wander}
\title{Wander}
\usage{
wander(
  n_lines = 100,
  n_points = 350,
  r_outer = 8,
  r_inner = 3,
  line_var = 0.01,
  deg_jitter = 0.1,
  linewidth = 0.1,
  bg_col = "#462255",
  col_palette = c("#FF8811", "#9DD9D2", "#046E8F", "#D44D5C"),
  n_cols = 20,
  s = 123
)
}
\arguments{
\item{n_lines}{Number of lines. Default 100.}

\item{n_points}{Number of points. Default 350.}

\item{r_outer}{Radius of outer circle. Default 8.}

\item{r_inner}{Radius of inner circle. Default 3.}

\item{line_var}{Variance of random walk noise. Default 0.01.}

\item{deg_jitter}{Degree of jitter for multiple lines. Default 0.1.}

\item{linewidth}{Width of lines. Default 0.1.}

\item{bg_col}{Background colour. Default "#462255".}

\item{col_palette}{Vector of colours. Default \code{c("#FF8811", "#9DD9D2", "#046E8F", "#D44D5C")}}

\item{n_cols}{Number of colours to create. Default 20.}

\item{s}{Random seed. Default 123.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using random walks
}
