% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waves.R
\name{waves}
\alias{waves}
\title{Waves}
\usage{
waves(a = 23, b = 6, main_col = "black", bg_col = "white", s = 2021)
}
\arguments{
\item{a}{sine wave parameter. Default 23.}

\item{b}{Cosine wave parameter. Default 6.}

\item{main_col}{Colour of lines. Either a single colour or one of the colour palettes from rcartocolor. Default "black".}

\item{bg_col}{Background colour. Default "white".}

\item{s}{Seed value. Default 2021.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object from sine and cosine waves.
}
