% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_boxes.R
\name{window_boxes}
\alias{window_boxes}
\title{Window Boxes}
\usage{
window_boxes(
  n_x = 10,
  n_y = 10,
  col_palette = PrettyCols::prettycols("Beach", n = 5),
  linewidth = 2,
  rayshade = FALSE
)
}
\arguments{
\item{n_x}{Number of squares per row. Default 10.}

\item{n_y}{Number of squares per column. Default 10.}

\item{col_palette}{Vector of colours. Default
\code{PrettyCols::prettycols("Beach", n = 5)}.}

\item{linewidth}{Width of borders around squares. Default 2.}

\item{rayshade}{Boolean determining whether the returned plot should be
converted to three dimensional using rayshader. If \code{TRUE}, \code{{rayshader}}
is required to be installed. Default \code{FALSE}.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using rectangles.
}
\examples{
window_boxes()
}
