% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRt-package.R
\docType{package}
\name{aRt-package}
\alias{aRt}
\alias{aRt-package}
\title{aRt: Generative Art with R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Creates generative art using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nrennie.github.io/aRt/}
}

}
\author{
\strong{Maintainer}: Nicola Rennie \email{nrennie35@gmail.com}

}
\keyword{internal}
