% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRt-package.R
\docType{package}
\name{aRt-package}
\alias{aRt}
\alias{aRt-package}
\title{aRt: Generative Art with R}
\description{
Creates generative art using R.
}
\author{
\strong{Maintainer}: Nicola Rennie \email{nrennie35@gmail.com}

}
\keyword{internal}
