% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attraction.R
\name{attraction}
\alias{attraction}
\title{Attraction}
\usage{
attraction(
  n = 5000,
  a = -3,
  b = 1,
  c = 0.5,
  d = -1,
  main_col = "black",
  bg_col = "white"
)
}
\arguments{
\item{n}{Number of points. Default 5000.}

\item{a}{Parameter. Default -3.}

\item{b}{Parameter. Default 1.}

\item{c}{Parameter. Default 0.5.}

\item{d}{Parameter. Default -1.}

\item{main_col}{Vector of colours (or single colour). Default \code{"black"}.}

\item{bg_col}{Background colour. Default \code{"white"}.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object using strange attractors.
}
\examples{
attraction(n = 100)
}
