% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullseye.R
\name{bullseye}
\alias{bullseye}
\title{Bullseye}
\usage{
bullseye(main_col = "black", bg_col = "white", s = 1234)
}
\arguments{
\item{main_col}{Colour scheme of art. Default "black".}

\item{bg_col}{Background colour. Default "white".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a layered generative art ggplot object using polar
bar charts.
}
\examples{
bullseye()
}
