% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{contours}
\alias{contours}
\title{Contours}
\usage{
contours(
  xmin = -3.27,
  xmax = -3.15,
  ymin = 54.43,
  ymax = 54.49,
  col_palette = MetBrewer::met.brewer("Hokusai3"),
  res = 50,
  line_col = "#0a2e57",
  linewidth = 0.5
)
}
\arguments{
\item{xmin}{Minimum x value og geographic area. Default -3.27.}

\item{xmax}{Maximum x value og geographic area. Default -3.15.}

\item{ymin}{Minimum y value og geographic area. Default 54.43.}

\item{ymax}{Maximum y value og geographic area. Default 54.49.}

\item{col_palette}{Colour palette to colour in contours.
Default MetBrewer::met.brewer("Hokusai3").}

\item{res}{Resolution of elevation. Default 50.}

\item{line_col}{Colour of contours. Default "#0a2e57".}

\item{linewidth}{Width of contours. 0.5.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot
object based on elevation data of a geographic area
}
