% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Divide}
\usage{
divide(
  num_lines = 30,
  col_palette = PrettyCols::prettycols("TangerineBlues"),
  rayshade = FALSE,
  s = 1234
)
}
\arguments{
\item{num_lines}{Number of intersecting lines. Default \code{30}.}

\item{col_palette}{Vector of colours. Default \code{PrettyCols::prettycols("TangerineBlues")}.}

\item{rayshade}{Boolean determining whether the returned plot should be converted to
three dimensional using rayshader. If \code{TRUE}, \code{{rayshader}} is required to be installed.
Default \code{FALSE}.}

\item{s}{Seed value. Default \code{1234}.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object from intersecting lines.
}
