% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fading.R
\name{fading}
\alias{fading}
\title{Fading}
\usage{
fading(
  n_layers = 6,
  n_points = 10,
  col_palette = rcartocolor::carto_pal(n = 7, "SunsetDark"),
  s = 1234
)
}
\arguments{
\item{n_layers}{Number of layers. Default 6.}

\item{n_points}{Number of points per layer area. Default 10.}

\item{col_palette}{Vector of colours. Default "SunsetDark" colour palette from rcartocolor.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using voronoi tiles.
}
