% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients.R
\name{gradients}
\alias{gradients}
\title{Generates generative art as a grid of gradient colour fades}
\usage{
gradients(nx = 5, ny = 5, s = 1234)
}
\arguments{
\item{nx}{Number of columns. Default 5.}

\item{ny}{Number of rows. Default 5.}

\item{s}{Random seed. Default 1234.}
}
\value{
Tibble
}
\description{
Generates generative art as a grid of gradient colour fades
}
\examples{
gradients()
}
