% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrored.R
\name{mirrored}
\alias{mirrored}
\title{Mirrored}
\usage{
mirrored(
  n = 15,
  w = 4,
  col_palette = PrettyCols::prettycols("PurpleTangerines"),
  s = 1234
)
}
\arguments{
\item{n}{Number of boxes per quadrant. Default 15.}

\item{w}{Weighting towards first colour of palette. Minimum of 2. Default 4.}

\item{col_palette}{Vector of colours. Default PrettyCols::prettycols("PurpleTangerines").}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object from rectangles.
}
