% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{rectangles}
\alias{rectangles}
\title{Rectangles}
\usage{
rectangles(
  n = 100,
  max_height = 7,
  max_width = 5,
  size = 2,
  main_col = "lightgrey",
  col_palette = "Bold",
  bg_col = "white",
  s = 123
)
}
\arguments{
\item{n}{Number of rectangles. Default 100.}

\item{max_height}{Maximum height of rectangle. Default 7.}

\item{max_width}{Maximum width of rectangle. Default 5.}

\item{size}{Line width of rectangles. Default 2.}

\item{main_col}{Colour of non-highlighted rectangles. Default "lightgrey".}

\item{col_palette}{Colour palette from rcartocolor for highlighted rectangles. Must have 12 colours.}

\item{bg_col}{Background colour. Default "white".}

\item{s}{Seed value. Default 123.}
}
\value{
A ggplot object
}
\description{
This function generates a generative art ggplot object featuring multiple coloured rectangles.
}
