% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro.R
\name{spiro}
\alias{spiro}
\title{Spiro}
\usage{
spiro(
  n_x = 10,
  n_y = 10,
  d = 10,
  R = 4,
  r = 1,
  linewidth = 0.5,
  col_palette = "white",
  bg_col = "grey20",
  s = 1234
)
}
\arguments{
\item{n_x}{Number of spirals per row. Default 10.}

\item{n_y}{Number of spirals per column. Default 10.}

\item{d}{Diameter. Default 10.}

\item{R}{Outer radius. Default 4.}

\item{r}{Inner radius. Default 1.}

\item{linewidth}{Width on lines. Default 0.5.}

\item{col_palette}{Vector of colours. Default "white".}

\item{bg_col}{Background colour. Default "grey20".}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object
}
\description{
This function generates a generative art ggplot object in a spirograph style.
}
\examples{
spiro()
}
