% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static}
\alias{static}
\title{Generate Static Art}
\usage{
static(perc, s = 1234)
}
\arguments{
\item{perc}{Percentage of data points to be non-NA}

\item{s}{Seed value}
}
\value{
A ggplot object
}
\description{
This function generates a greyscale generative art ggplot object.
}
