% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripes.R
\name{stripes}
\alias{stripes}
\title{Stripes}
\usage{
stripes(
  perc = 0.5,
  n = 3,
  col_palette = c("#8EA604", "#F5BB00", "#EC9F05", "#D76A03", "#BF3100"),
  alpha = 1,
  s = 1234
)
}
\arguments{
\item{perc}{Percentage of data points to be sorted. Default 0.5.}

\item{n}{Number of rows. Default 3.}

\item{col_palette}{Vector of colours. Default \code{c("#8EA604", "#F5BB00", "#EC9F05", "#D76A03", "#BF3100")}.}

\item{alpha}{Transparency of fill. Default 1.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object
}
\description{
This function generates a generative art ggplot object featuring rows of stripes.
}
\examples{
stripes()
}
