% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunsets.R
\name{sunsets}
\alias{sunsets}
\title{Sunsets}
\usage{
sunsets(
  num_bars = 8,
  n = 1000,
  col_palette = PrettyCols::prettycols("Lively"),
  bg_col = "#413C58",
  vertical = FALSE,
  fade_vertical = FALSE,
  alpha = 1,
  s = 2023
)
}
\arguments{
\item{num_bars}{Number of bars. Default 8.}

\item{n}{Number of lines per bar. Default 1000.}

\item{col_palette}{Colour palette. Default PrettyCols::prettycols("Lively").}

\item{bg_col}{Background colour. Default "#413C58".}

\item{vertical}{Boolean indicating whether bars should be
vertical. Default FALSE.}

\item{fade_vertical}{Boolean indicating whether the colouring
should be vertical. Default FALSE.}

\item{alpha}{Transparency of coloured bars. Default 1.}

\item{s}{Random seed. Default 2023.}
}
\value{
A ggplot object.
}
\description{
This function creates generative art using faceted segments.
}
