% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{tiles}
\alias{tiles}
\title{Tiles}
\usage{
tiles(n_x = 12, n_y = 12, col_palette = "Veronese", num_colours = 5, s = 1234)
}
\arguments{
\item{n_x}{Number of polygons per row. Default 12.}

\item{n_y}{Number of polygons per column. Default 18.}

\item{col_palette}{Colour palette from MetBrewer.}

\item{num_colours}{Number of colours to use.}

\item{s}{Seed value. Default 1234.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using square polygons.
}
