% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{windows}
\alias{windows}
\title{Windows}
\usage{
windows(
  n_x = 10,
  n_y = 10,
  col_palette = PrettyCols::prettycols("Beach", n = 5),
  linewidth = 2
)
}
\arguments{
\item{n_x}{Number of squares per row. Default 10.}

\item{n_y}{Number of squares per column. Default 10.}

\item{col_palette}{Vector of colours. Default PrettyCols::prettycols("Beach", n = 5).}

\item{linewidth}{Width of borders around squares. Default 2.}
}
\value{
A ggplot object.
}
\description{
This function generates a coloured generative art ggplot object using rectangles.
}
